IDP-877 Better handling of band configuration
https://issues.shibboleth.net/jira/browse/IDP-877

If the user specified a dependency which is a dataconnector, but no sourceAttributeID we issue an error message and return no values (resulting in the attribute being resolved empty and thus thrown away.

The error is warned on in the parser, but this change stops the resolver throwing an uncaught exception and the end user into an error page.